requirements.txt
aiohttp==3.10.11
httpx==0.27.2
tenacity==9.0.0
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.13
redis==5.1.1
alembic==1.13.3
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.5.2
python-telegram-bot==21.7
APScheduler==3.10.4
orjson==3.10.11
